# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EcoLearn API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  # ─── Dashboard ───────────────────────────────────────────
...
  /lessons:
    get:
      operationId: listLessons
      tags: [lessons]
      summary: List all lessons
      parameters:
        - name: topic
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"
    post:
      operationId: createLesson
      tags: [lessons]
      summary: Create a lesson (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonInput"
      responses:
        "201":
          description: Created lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /lessons/{id}:
    get:
      operationId: getLesson
      tags: [lessons]
      summary: Get a lesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
        "404":
          description: Not found
    patch:
      operationId: updateLesson
      tags: [lessons]
      summary: Update a lesson (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonUpdate"
      responses:
        "200":
          description: Updated lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
    delete:
      operationId: deleteLesson
      tags: [lessons]
      summary: Delete a lesson (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Quizzes ─────────────────────────────────────────────
...
  /challenges:
    get:
      operationId: listChallenges
      tags: [challenges]
      summary: List all eco challenges (with completion status for current user)
      responses:
        "200":
          description: List of challenges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChallengeWithStatus"
    post:
      operationId: createChallenge
      tags: [challenges]
      summary: Create an eco challenge (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChallengeInput"
      responses:
        "201":
          description: Created challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Challenge"
  /challenges/{id}:
    patch:
      operationId: updateChallenge
      tags: [challenges]
      summary: Update an eco challenge (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChallengeUpdate"
      responses:
        "200":
          description: Updated challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Challenge"
    delete:
      operationId: deleteChallenge
      tags: [challenges]
      summary: Delete an eco challenge (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /challenges/{id}/complete:
    post:
      operationId: completeChallenge
      tags: [challenges]
      summary: Mark a challenge as completed by the current student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completion recorded with points awarded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeCompletion"

  # ─── Leaderboard ─────────────────────────────────────────
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password, school]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        school:
          type: string
        grade:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, email, role, ecoPoints]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, admin]
        ecoPoints:
          type: integer
        school:
          type: ["string", "null"]
        grade:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Lesson:
      type: object
      required: [id, title, topic, summary, content, imageUrl, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        topic:
          type: string
          enum: [climate_change, waste_management, water_conservation, energy_saving, biodiversity]
        summary:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        createdAt:
          type: string
          format: date-time
    LessonInput:
      type: object
      required: [title, topic, summary, content, imageUrl]
      properties:
        title:
          type: string
        topic:
          type: string
          enum: [climate_change, waste_management, water_conservation, energy_saving, biodiversity]
        summary:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
    LessonUpdate:
      type: object
      properties:
        title:
          type: string
        topic:
          type: string
          enum: [climate_change, waste_management, water_conservation, energy_saving, biodiversity]
        summary:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]

    # ─── Quizzes ─────────────────────────────────────────────
...
    Challenge:
      type: object
      required: [id, title, description, icon, ecoPoints, isDaily, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        ecoPoints:
          type: integer
        isDaily:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ChallengeWithStatus:
      type: object
      required: [id, title, description, icon, ecoPoints, isDaily, completedToday, totalCompletions, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        ecoPoints:
          type: integer
        isDaily:
          type: boolean
        completedToday:
          type: boolean
        totalCompletions:
          type: integer
        createdAt:
          type: string
          format: date-time
    ChallengeInput:
      type: object
      required: [title, description, icon, ecoPoints]
      properties:
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        ecoPoints:
          type: integer
        isDaily:
          type: boolean
    ChallengeUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        ecoPoints:
          type: integer
        isDaily:
          type: boolean
    ChallengeCompletion:
      type: object
      required: [id, challengeId, pointsAwarded, totalPoints, newBadges]
      properties:
        id:
          type: integer
        challengeId:
          type: integer
        pointsAwarded:
          type: integer
        totalPoints:
          type: integer
        newBadges:
          type: array
          items:
            $ref: "#/components/schemas/Badge"

    # ─── Leaderboard ─────────────────────────────────────────
...
    Badge:
      type: object
      required: [id, name, description, icon, pointsRequired]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        pointsRequired:
          type: integer
        earnedAt:
          type: ["string", "null"]