# lib/api-zod/src/generated/types/challenge.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EcoLearn API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Challenge {
  id: number;
  title: string;
  description: string;
  icon: string;
  ecoPoints: number;
  isDaily: boolean;
  createdAt: Date;
}