# lib/api-zod/src/generated/types/challengeCompletion.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EcoLearn API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Badge } from './badge';

export interface ChallengeCompletion {
  id: number;
  challengeId: number;
  pointsAwarded: number;
  totalPoints: number;
  newBadges: Badge[];
}