# lib/api-zod/src/generated/types/challengeInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EcoLearn API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChallengeInput {
  title: string;
  description: string;
  icon: string;
  ecoPoints: number;
  isDaily?: boolean;
}