# lib/api-zod/src/generated/types/lessonInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EcoLearn API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LessonInputDifficulty } from './lessonInputDifficulty';
import type { LessonInputTopic } from './lessonInputTopic';

export interface LessonInput {
  title: string;
  topic: LessonInputTopic;
  summary: string;
  content: string;
  imageUrl: string;
  difficulty?: LessonInputDifficulty;
}