# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EcoLearn API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  ecoPoints: number;
  /** @nullable */
  school?: string | null;
  /** @nullable */
  grade?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt?: Date;
}